+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed generating any-type schemas [#3536](https://github.com/microsoft/kiota/issues/3536)
-- Fixed float to string convertation with invariant culture. 
+- Fixed float to string conversion with invariant culture.
+- Fixed a bug where vendor mime types would be ignored. [#3666](https://github.com/microsoft/kiota/issues/3666)
 
 ## [1.8.1] - 2023-11-02